+++ packages/components-react/CHANGELOG.md
@@ -3,6 +3,30 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [5.1.0-next.30](https://github.com/Esri/calcite-design-system/compare/@esri/calcite-components-react@5.1.0-next.29...@esri/calcite-components-react@5.1.0-next.30) (2026-04-07)
+
+**Note:** Version bump only for package @esri/calcite-components-react
+
+## [5.1.0-next.29](https://github.com/Esri/calcite-design-system/compare/@esri/calcite-components-react@5.1.0-next.28...@esri/calcite-components-react@5.1.0-next.29) (2026-04-07)
+
+**Note:** Version bump only for package @esri/calcite-components-react
+
+## [5.1.0-next.28](https://github.com/Esri/calcite-design-system/compare/@esri/calcite-components-react@5.1.0-next.27...@esri/calcite-components-react@5.1.0-next.28) (2026-04-06)
+
+**Note:** Version bump only for package @esri/calcite-components-react
+
+## [5.1.0-next.27](https://github.com/Esri/calcite-design-system/compare/@esri/calcite-components-react@5.1.0-next.26...@esri/calcite-components-react@5.1.0-next.27) (2026-04-03)
+
+**Note:** Version bump only for package @esri/calcite-components-react
+
+## [5.1.0-next.26](https://github.com/Esri/calcite-design-system/compare/@esri/calcite-components-react@5.1.0-next.25...@esri/calcite-components-react@5.1.0-next.26) (2026-03-31)
+
+**Note:** Version bump only for package @esri/calcite-components-react
+
+## [5.1.0-next.25](https://github.com/Esri/calcite-design-system/compare/@esri/calcite-components-react@5.1.0-next.24...@esri/calcite-components-react@5.1.0-next.25) (2026-03-31)
+
+**Note:** Version bump only for package @esri/calcite-components-react
+
 ## [5.1.0-next.24](https://github.com/Esri/calcite-design-system/compare/@esri/calcite-components-react@5.1.0-next.23...@esri/calcite-components-react@5.1.0-next.24) (2026-03-27)
 
 **Note:** Version bump only for package @esri/calcite-components-react